Fix null strings
This commit fixes null strings to support types with no namespace. Also, user strings and blobs are supported with null references now (provides a null metadata handle instead of throwing an exception). A unit test is provided with a anonymous type and a type with no namespace to support this commit.

See: #9, #28, #33