Fix inline constructor reference resolving
In old implementation, if a type reference is found, all of its public constructor references are resolved. But, if a type reference is not referenced as a type at anywhere and one of its constructor is invoked in a method body, old implementation mistakenly assumed constructor reference should be available. This commit fixes this problem. Also, a new unit test is added to demonstrate the problem.

See: #33