Fix for AssemblyGeneratorTests randomly failing
When running under Visual Studio the tests in `AssemblyGeneratorTest`
class often fail with file access errors when trying to save the
generated assembly. Not sure exactly why but I suspect Visual Studio
might be watching this folder for changes and load modified DLL's
causing the write conflict.

Putting all the output files in a `/generated` sub-folder seems to fix it.

Lokad/ILPack#63